Move old test doubles into new location
