+++ tests/TestDoubles/GenericResourcePrimary.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Procountor\Tests;
+namespace Procountor\Tests\TestDoubles;
 
 use DateTime;
 use Procountor\Procountor\Collection\AbstractCollection;
 use Procountor\Procountor\Interfaces\AbstractResourceInterface;
 
 
-class TestResourcePrimary implements AbstractResourceInterface
+class GenericResourcePrimary implements AbstractResourceInterface
 {
 
     public function getTestDate(): DateTime
@@ -22,7 +22,7 @@ class TestResourcePrimary implements AbstractResourceInterface
 
     public function getTestAnotherResource(): ?AbstractResourceInterface
     {
-        return new TestResourceSecondary();
+        return new GenericResourceSecondary();
     }
 
     public function getTestString(): string
@@ -37,7 +37,7 @@ class TestResourcePrimary implements AbstractResourceInterface
 
     public function getTestCollection(): AbstractCollection
     {
-        return new TestCollection(...array_fill(0, 5, new TestResourceSecondary()));
+        return new Collection(...array_fill(0, 5, new GenericResourceSecondary()));
     }
 
 }
