+++ tests/Unit/BuilderTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use Procountor\Procountor\Json\Builder;
-use Procountor\Tests\TestResourcePrimary;
+use Procountor\Tests\TestDoubles\GenericResourcePrimary;
 use Procountor\Tests\UnitTestCase;
 
 test('build json from TestResource', function () {
@@ -9,7 +9,7 @@ test('build json from TestResource', function () {
     /** @var UnitTestCase $this */
 
     $builder = new Builder();
-    $builder->setResource(new TestResourcePrimary());
+    $builder->setResource(new GenericResourcePrimary());
 
     // Parse & compare to avoid annoying differencies in whitespace etc.
     $actual = json_decode($builder->getJson());
