+++ tests/Unit/CollectionTest.php
@@ -1,6 +1,7 @@
 <?php
 
-use Procountor\Tests\TestCollection;
+use Procountor\Tests\TestDoubles\Collection;
+use Procountor\Tests\TestDoubles\GenericResourcePrimary;
 use Procountor\Tests\TestResourcePrimary;
 use Procountor\Tests\UnitTestCase;
 
@@ -9,7 +10,7 @@ test('adding invalid item throws', function () {
 
     /** @var UnitTestCase $this */
 
-    $collection = new TestCollection();
+    $collection = new Collection();
     $collection->addItem(new stdClass());
 
     $this->assertCount(0, $collection);
@@ -21,8 +22,8 @@ test('adding valid item', function () {
 
     /** @var UnitTestCase $this */
 
-    $collection = new TestCollection();
-    $collection->addItem(new TestResourcePrimary());
+    $collection = new Collection();
+    $collection->addItem(new GenericResourcePrimary());
 
     $this->assertCount(1, $collection);
 
@@ -33,7 +34,7 @@ test('iterating over collection', function() {
 
     /** @var UnitTestCase $this */
 
-    $collection = new TestCollection(...array_fill(0, 5, new TestResourcePrimary()));
+    $collection = new Collection(...array_fill(0, 5, new GenericResourcePrimary()));
 
     $this->assertCount(5, $collection);
